Resurrect the SIGILL detection method for dotprod instructions, because recent Android devices supporting dotprod instructions are still using kernels not supporting the auxv detection method.
Also, our linux version check was incorrect in that it was based on Pixel4 device trees, but other devices have not received the same backports in the same kernel versions, so instead we now require the linux version, 4.14.151, where this was backported to the mainline linux tree.

Another tweak is that we use std::tuple::operator>= for the version check.

This is essentially a rollback of git commit 23adc555, piper cl/288340160.

PiperOrigin-RevId: 311663777